requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
appdirs==1.4.4
async-timeout==5.0.1
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
bigdl-chronos-spark3==2.5.0b20240324
bigdl-core==2.4.0.dev0
bigdl-dllib-spark3==2.5.0b20240324
bigdl-friesian-spark3==2.5.0b20240324
bigdl-math==2.4.0.dev0
bigdl-nano==2.5.0b20240324
bigdl-orca-spark3==2.5.0b20240324
bigdl-serving==2.5.0b20240324
bigdl-spark3==2.5.0b20240324
bigdl-tf==2.4.0.dev0
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
conda-pack==0.3.1
configobj==5.0.9
contourpy==1.3.0
cryptography==45.0.7
cycler==0.12.1
databricks-sdk==0.64.0
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==0.24.1
dvc==3.63.0
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
Flask==3.1.2
flatbuffers==25.2.10
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2024.6.1
funcy==2.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
gto==1.8.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imgviz==1.7.6
importlib_metadata==8.7.0
importlib_resources==6.5.2
intel-cmplr-lib-ur==2025.2.1
intel-openmp==2025.2.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.5.2
kafka-python==2.2.15
kiwisolver==1.4.7
kombu==5.5.4
labelme==5.8.3
lazy_loader==0.4
loguru==0.7.3
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
mlflow==3.1.4
mlflow-skinny==3.1.4
mpmath==1.3.0
multidict==6.6.4
natsort==8.4.0
networkx==3.2.1
numpy
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
onnxruntime==1.16.3
opencv-python==4.11.0.86
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
osam==0.2.5
osrf-pycommon==2.0.0
packaging==25.0
pandas
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf
psutil==7.0.0
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pygit2==1.15.1
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PySocks==1.7.1
pyspark==3.4.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
scikit-image==0.24.0
scikit-learn==1.0.2
scipy==1.13.1
scmrepo==3.5.2
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
sigfig==1.3.19
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
sqltrie==0.11.2
starlette==0.47.3
sympy==1.14.0
tabulate==0.9.0
tcmlib==1.4.0
threadpoolctl==3.6.0
tifffile==2024.8.30
tomli==2.2.1
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
umf==0.11.0
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.20.1
zc.lockfile==3.0.post1
zipp==3.23.0
